fix(bot-client): break five circular dependencies in deny/ and character/
Both cycles had accumulated on develop for weeks and were only caught by a
fresh depcruise run. Structural fix rather than suppression:

deny/ (4 cycles): extract DenylistEntryResponse into a new leaf
browseTypes.ts and move the registerBrowseRebuilder call into its own
browseRebuilder.ts. detail.ts and friends now import the shared type
from the leaf, and the rebuilder registration is triggered via a
dedicated side-effect import in index.ts. Removes the
browse.ts ↔ detail.ts edges (dynamic import + side-effect import)
that depcruise was flagging.

character/ (1 cycle): extract CharacterData into a new leaf
characterTypes.ts. sections.ts now imports the type from the leaf
rather than from config.ts, which in turn imports sections.ts to
compose the dashboard config. config.ts keeps a type re-export so
the 13+ other callers stay unaffected.

Test split: the two rebuilder callback tests move out of browse.test.ts
into the new colocated browseRebuilder.test.ts, matching the source
split (browseRebuilder.ts is a logic module and needs its own test).
